Add parameter value support to callback function.
We pass all parameters of a called function to the callback
function by passing an array of pointers to their position
on the stack.

In combination with the argument types we can reconstruct the
parameter values for IntegerTypes at run time and generate
new function versions based on these values.