Enable PAPI CScopProfiling.
We provide basic support for PAPI CScop profiling for now.
Functionality is shared with PapiRegionProfiling.

We do not require a PapiRegionPrepare pass. Instead of trying
to make that abomination work again, we should rely
on the FunctionCloner of polli instead.

In order to instrument without recompilation use the following flags
for polli:
  -instrument -caddy -no-recompilation

Do not forget to add the necessary libraries:
  -lpapi -lpprof

Support for building libpprof.so is required in the pprof-study.