Tracing support.
We can now monitor arbitrary code regions (inside JIT / inside JIT'ed
code) with an (almost) arbitrary set of PAPI counters.

First you have to add the events you want to monitor during this
session, once with pprof_add_event(...).
Commit the setup using pprof_trace_start(). This starts the event
set as you configured it.

In your code mark regions of interest with pprof_trace_entry(<name>)
and pprof_trace_exit(<name>).

In the end, finish the collection with pprof_trace_stop(). This stores
the collected events to disk in: ./papi.profile.out

This patch misses the analysis support for these types of profiles. It
follows in the next patch.