pjit: Introduce new type RuntimeValues
This replaces the ancient ParamVector types with a new RuntimeValue type. We use
this to wrap alle runtime values that get handed to the JIT. All values are
stored inside the RunValueList, which can hash all values that are used for
specialization.

The key functions that used Param & ParamVector types now use the new RunValue
types.

This, interestingly, fixes at least one SIGTRAP in x264.