pgsql: delete simple run after processing.
Let's hope we don't run into errors along the way. The transaction
will finish after read & deleting the run.

We will lose all information about the run, if we crash in a later stage,
before the events are aggregated and written back to the db.

This needs to change in a later commit.