add support for regression test collection
This is the JIT-side commit that is required for automatic regression test
generation. The JIT will persist all extracted modules in the pprof database.

This commit introduces a libPolyJIT_Persist, which will handle all databse-
related queries by introducing a minimalistic Object-Relational manager.
So far, this only includes a Session object (aka Transaction) and
a RegressionTest schema object.

In addition to this, we introduce the ability to add UnitTests for all polli
related commits.