pjit: from sequential to parallel code generation
This is the first step for PolyJIT to become a parallel JIT.
All sequential phases are now delegated into a new thread 'CodeGen'
that performs all the work from:
* Analyzing the run-time parameters.
* Creating new variants
* Preparing the new object code in memory.

From now on pjit_main just triggers a new SpecializerRequest that
contains all necessary information for PolyJIT to continue.
In this first step pjit_main will wait (blocking) for the new
variant to be available in the FunctionCache before execution continues.

In a next step we can go to (full-)asynchronous compilation by falling
back to the sequential version, while CodeGen finishes. This will calibrate
the JIT automatically and reduce our perceived overhead in total.