pjit_main: return boolean from jit callback
This changes the return type of pjit_main from void
to bool. If we return true, we want to exit like before.
In a later commit we will jump to the sequential code
as soon as we return false.
This will enable auto-calibration of variant generation against
the code generator.

In this commit we encounter 'unreachable' as soon as we would return
false from pjit_main. All relevant tests are updated to reflect
this change.