pjit: copy runtime parameters when a new request is generated
This fixes a race-condition between the CodeGen thread and the main
thread of the program:

Before this commit it is possible that the sequential execution of the
function is finished before CodeGen can generate a variant. This causes
the wrapper function to return, which cleans the stack with our runtime
parameters array.

This commit copies the run-time parameters into an array that is managed
by the SpecializerRequest.