Heuristic: Block the JIT only the first time we encounter a SCoP.
This commit blocks the JIT only for the first time we encounter a new
SCoP function. This will enable optimization of SCoPs that only get
called once.
If SCoPs get called more than once, we want to make sure that we catch
those that have enough run-time that the code-generator can catch up
with the repeated calls.