pjit: rework asynchronous compilation
This huge commit does the following (hard to pull the bits apart after this
day).

1) Use asynchronous tasks with the introduced TaskSystem to queue code
generation tasks.
2) Generate all variants against a single execution engine. This will only
compile the added modules.
Optimize: Make sure the engine looks at the new elements first, when
finalizeObject() is called.
3) Move everything that requires initialization into the PolyJIT class. This way
we only have one place to look at when dealing with the
intiialization/deinitialization phase. This is usefull when we need to cancel
all pending codegen tasks.
4) Reduce interface clutter with VariantFunctions/Dispatcher and so on. Lots of
useless stuff in there. Probably more.