extract: tag the Values the JIT needs to specialize at run time.
With this commit, polli will be able to tag the function arguments it needs
to specialize on. First we extract the set of required parameters of a JIT SCoP.
Afterwards we find all corresponding LLVM Values that will be on the incoming
edge of the SCoP region.
Finally, all Values that lie on an incoming edge will be matched with the
functions arguments of the extracted function. All matches will be tagged
with the attribute 'polli.specialize'.

Missing parts:
* Make sure the Function cloner copies the attributes (untested).
* Read the attribute at run time and act accordingly.