extract: modify pointer operands instead of replacing them
This commit replaces pointer operands of instructions that contain
ConstantExpressions. Before, we would create a new Instruction with
the ConstantExpr part replaced by an Instruction Value and replace
the old Instruction with the new one.
This triggered various errors with AssertingValueHandles, therefore
we now just modify the existing (working) code.