extract: prepare regions for extraction
This provides a bug in llvm's CodeExtractor. In the case of a region having more
than
one exiting edge, the CodeExtractor fails to preserve the dominance information
properly.

We fix it by providing a single exit edge for the region. A simple split of the
exit's predecessors that
belong to the region solves this particular problem.