libpjit: lookup global variables from prototypes inside the binary
First we load all symbols from the binary with dlload. Afterwards we try
to find all GlobalVariables from the Prototype with dlsym and setup a
mapping from GlobalVariale to Address.

For this to work you need to make sure to have a Dynamic Symbol Table inside
the ELF object. Use -rdynamic to get it.