cloner: do not 'share' internal linkage symbols
This is an experimental commit. It may break everything...

We try to leave internal symbols alone. We cannot find them at run time, because
the linker does not see them anymore.

If we break stuff, the natural way of handling internal linkage global values is
to
reintroduce the "Redirect Global Values as Function Parameters" feature we had
in
PolyJIT a few months ago.