cloner: deal with globals  with internal linkage
This commit teaches the cloner to deal with global values that have
internal linkage.
We rename the original global variable by giving it the name of the target
module of the
cloner as suffix. Afterwards we change the linkage from InternalLinkage to
ExternalLinkage.

This makes the symbol visible to the linker at run time without causing a
name-collision.
Further clones in the same module will not hurt, as the symbol already has
ExternalLinkage.