Split PolyJIT's runtime pipeline into two parts.
1: The default 'RELEASE' pipeline that should be used when you want to
generate code in a release environment without interruption by
diagnostics.

2: The 'DEBUG' pipeline that is enabled when you want to see
additional diagnostic information.

You can either enable the pipeline via command-line manually using the
switch
    -polli-optimizer=(release,debug)

or enable the debug pipeline implicitly as soon as you use one of the
new configuration options:
    -polli-database-export
    -polli-scheudle-report
    -polli-scop-report
    -polli-ast-report
    -polli-function-report

This commit does not finish the separation completely. There is still a
lot of logging output going on that needs to be tamed. Following commits
will deal with that.