extract: increase the region depending on the required parameters.
This commit tries to find a bigger region than the detected SCoP one to
extract. During extraction we identify the parameter values that we need
for the SCoP at run-time. However, the definition of those might be
further up in the region tree.

We now have 2 possible scenarios:
  1) We extract the biggest region that defines one of our tracked
  parameters.
  2) We extract the smallest region that defines one of our tracked
  parameters.

For now we try out option 1). Option 2) seems to be good in the case
where they are not all defined on the same block (Risk: more
casts/operations interfering with our specialization).