jit: thread-safety part 2.
We still had lots of problems when LLVM was running in multi-threaded
mode. One of the main causes was the SpecialzerRequest data strucutre.

With this commit lots of small changed are introduced to the jit that
are not further mentioned here. Most revolve around small type-fixes in
function signatures, logging output and general compatibility with
non-debug builds of LLVM.