[FIX] libpjit: enforce construction/destruction order.
In some cases we could run into the problem where we would free all
LLVM modules but are still compiling inside the thread pool because the
Compiler object got destroyed while the thread pool was not yet
finished.

With this commit we factorize the thread pool out of the jit context.
This will ensure (in the current implementation of pjit_main) that the
ManagedStatic objects of LLVM are initialized in the order:
  1. JitContext
  2. Compiler
  3. ThreadPool

Destruction follows the reverse order and should guarantee that the
thread pool is finished before we touch any LLVM modules.