extract: temporarily disable module verification
We got a lot of false positives from this verification and I am yet unsure why.
For now, I will disable the verification after extraction of a prototype module
and provide a late verification at run-time.

This will fail late and hard, if the problem is not a false positive.