Fix: Function Cloner does not copy functions that are marked "mapped"
This fixes a bug where a remapped function was not picked up by the cloner
because it is marked with the mapped_fn attribute.
As a result, cloning a prototype that includes calls to mapped functions results
in an error.

This commit remaps the function again, if it is not present in the module,
regardless of the attribute.